+++ purgecss.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  content: ['index.html', 'src/**/*.vue'],
+  css: ['src/css/main.css']
+};